Remove extra interface down
This code was put here because WifiStateMachine#setupDriverForSoftAp()
had a similar clause.  However that code was a no-op, because the HAL
was stopped at the point that it was called, preventing WSM from getting
the interface index to turn down.

The very next thing that WSM did after setting the interface status to
down was to start the HAL, which sets the interface state to up anyway.

Change-Id: I184166bfd45b33171c3b2603a4ac3423e1674132