Add a copy counters for netlink packet debugging
Copy operation of NL80211Packet copies the entire buffer. This
costs a lot resouces. We should avoid it as much as we can.

This CL adds a static variable in NL80211Packet tracking the number
of copy operations, as well as another counter tracking the bytes of
we copied. This help us debug wificond netlink performance issues.

BUG=30901326
TEST=manual tests

Change-Id: I5be04ff977b5ffd7340a98d435a5703b0b79be28