Better logging for GetWiphyInfo()
NL80211_ATTR_WIPHY_BANDS contains a list of anonymous attributes.
Each anonymous attribute represents a band.
NL80211_BAND_ATTR_FREQ is expected to be included in every anonymous
attribute to specifying supported channels for this band.

However, sometimes one of the anonymous attribute provided by kernel is
empty, where NL80211_BAND_ATTR_FREQ is not included. This always happens
on some type of devices. We should not log an error for this.

Bug: 35220903
Test: compile, unit tests, manual test

Change-Id: I0ef953266744713af627a17e7d10f04b937e0648