Restart wificond on ENODEV scan failure
ENODEV scan failure implies that driver entered a bad state.
In order to recover wificond, we should treat it as an fatal
error. The crashing of wificond will trigger the restarting
of framework, bringing everything back to work.

Bug: 36894921
Test: compile, unit tests
Test: manual test:
Use "iw wlan0 del" to remove kernel interface.
Observe that wificond restarts and wifi continues to work.

Change-Id: I89f4403bcb570d7f413fa4cc9ca50378de428d3d