Wificond: Make Offload HAL service nonblocking
Offload HAL getService() is a blocking call which can cause a deadlock.
Replace it with non blocking call tryGetService() which returns null if
the service is not running instead of waiting on the service to start.
This implies that the client needs to be able to get a reference to the
service at a later time. The design choice is to use tryGetService() API
to get the handle when a disconnected mode PNO scan is requested if this
was not available before.

Bug: 62807929
Test: Unit test, sanity test
Change-Id: I344fd93c4afe2381c356e626e83311af49a5ad49