wificond: Move get*Channels() APIs
These get*Channels() APIs should be used by java framework
in both client and AP mode.

They are implemented by GetWiphy command which is associated to
a wihpy instead of a interface.

We need to move these APIs to wificond main interface for both
correctness and convienience.

This also removes the corresponding integration test. Since wificond
no longer load/unload driver/fw, we are not able to run and verify
these tests. It's reasonble to remove them for now.

Bug: 35707768
Test: compile, unit tests
Change-Id: I7ba562fbb788a6032832a0ab1ff917b5ba869670